chalenge 3 et modifier chalenge 2

+++ CHALENGE2.c
@@ -3,11 +3,11 @@
 
 int main()
 {
-    int fahrenheit;
+    float fahrenheit;
     printf("ENTRER LA TEMPERATURE EN FAHRENHEIT : ");
-    scanf("%d",&fahrenheit);
-    int result = (fahrenheit-32)* 5/9 ;
-    printf( "%d F= %d c",fahrenheit,result );
+    scanf("%f",&fahrenheit);
+    float result = (fahrenheit-32)* 5/9 ;
+    printf( "%f F= %f c",fahrenheit,result );
 
     return 0;
 }
